fix: DDL-Validate 통과하도록 column 자료형 수정

+++ src/main/java/com/swp/board/domain/Board.java
@@ -13,7 +13,7 @@ public class Board extends CreatedAtEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "board_id", nullable = false)
-    private Long boardId;
+    private Integer boardId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
